pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ircawp"
version = "0.1.0"
description = "Slack bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "aiohttp>=3.12.15",
    "aiosignal>=1.4.0",
    "appier>=1.34.8",
    "async-timeout>=5.0.1",
    "attrs>=25.3.0",
    "autopep8>=2.3.2",
    "beautifulsoup4>=4.13.5",
    "bitsandbytes>=0.48.2",
    "certifi>=2025.8.3",
    "charset-normalizer>=3.4.3",
    "commonmark>=0.9.1",
    "diffusers",
    "feedparser>=6.0.11",
    "frozenlist>=1.7.0",
    "idna>=3.10",
    "langchain>=0.3.27",
    "llama-cpp-python>=0.3.16",
    "multidict>=6.6.4",
    "numpy>=1.26.4",
    "optimum-quanto>=0.2.7",
    "packaging>=25.0",
    "playwright>=1.56.0",
    "pycodestyle>=2.14.0",
    "pygments>=2.19.2",
    "pyparsing>=3.2.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "slack-bolt>=1.24.0",
    "tomli>=2.2.1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "typing-extensions>=4.15.0",
    "urllib3>=2.5.0",
    "wikitextparser>=0.56.4",
    "yarl>=1.20.1",
]

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers" }
[tool.hatch.build.targets.wheel]
packages = [
    "app",
]
include = [
    "app/**",
]
exclude = [
    "**/__pycache__/**",
    "tests/**",
    "_archive/**",
    "attic/**",
]

[tool.hatch.build.targets.sdist]
include = [
    "app/**",
    "README.md",
    "pyproject.toml",
]
exclude = [
    "**/__pycache__/**",
    "tests/**",
    "_archive/**",
    "attic/**",
]
[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "plugin: Tests for plugin framework",
    "tools: Tests for tool framework",
]
